imors-backend/requirements.txt
asgiref==3.7.2
black==24.2.0
channels==4.0.0
click==8.1.7
Django==5.0.3
djangorestframework==3.14.0
flake8==7.0.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.10.0